+++ agent/api/src/main/java/com/walmartlabs/concord/agent/api/JobType.java
@@ -3,5 +3,6 @@ package com.walmartlabs.concord.agent.api;
 public enum JobType {
 
     JAR,
+    RUNNER,
     JUNIT_GROOVY
 }