+++ docker-images/agent/src/main/docker/start.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-python2 /opt/concord/agent/concord-agent-${project.version}/bin/launcher.py run
+APP_DIR="${concord.baseDir}/agent/concord-agent-${project.version}"
+
+export RUNNER_PATH="$APP_DIR/runner/concord-runner-${project.version}-jar-with-dependencies.jar"
+
+python2 ${APP_DIR}/bin/launcher.py run