+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ArchivePipeline.java
@@ -34,7 +34,6 @@ public class ArchivePipeline extends Chain {
                 injector.getInstance(DependenciesProcessor.class),
                 injector.getInstance(UserInfoProcessor.class),
                 injector.getInstance(RequestDataStoringProcessor.class),
-                injector.getInstance(RunnerProcessor.class),
                 injector.getInstance(LogFileProcessor.class),
                 injector.getInstance(ValidatingProcessor.class),
                 injector.getInstance(ArchivingProcessor.class),