+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ResumePipeline.java
@@ -14,7 +14,6 @@ public class ResumePipeline extends Chain {
         super(injector.getInstance(ResumeStateStoringProcessor.class),
                 injector.getInstance(RequestDataStoringProcessor.class),
                 injector.getInstance(LogFileProcessor.class),
-                injector.getInstance(RunnerProcessor.class),
                 injector.getInstance(ArchivingProcessor.class),
                 injector.getInstance(AgentProcessor.class),
                 injector.getInstance(StateTransferringProcessor.class),